Again a big batch of changes
In the PlotsModel, try to use more the abstractions it provides. We should
end up with not depending on having a ProxyModel at the view. It's not
good to have all the code checking for source/proxy for every index.

Make the Dimensions flags, so we can easily use it for filtering.