Simplify the process for starting the plotting
Reduce conversions on the expression: only change from equation to
operation once, then leave it like that.
Also this means that now implicit functions are not bool anymore.

To do that, it required to make some little changes in the registration
of the functions, luckily it was already abstracted, somewhat.