First upload of components for JIT support.
As a first approach we have:

JitAnalyzer: The heart of this feature, this class inherited from Analyzer but additionally can perform calculations with JIT.
LLVMIRExpressionWriter: This component transform the analitza tree into LLVM IR and thus let us perform JIT evaluations by using JitAnalyzer.
AnalitzaJitTest: Here we are going to test JitAnalyzer.
cmake/FindLLVM.cmake: we need this to know if the user have LLVM in his system.

Finally, this feature is optional for the user (at compilation time)