Add TypeCompiler and rename LLVMIRExpressionWriter to ValueCompiler
Rationale of TypeCompiler: In the future we could have a type system based on hierarchies,
so we can use here a visitor to convert the Analitza type into a valid LLVM IR type.
For now, this class will take some ExpressionType and will convert it into a IR type
using simple and direct methods like convertToIRType, etc. (the name of this methods
will be defined in the next commit)