Use the default RGB Colors for the XYZ Axis
Summary: Only using red to plot the axis is misleading as one cannot easily identify which axis is pointing to where. Using the default standard to use RGB helps on that. If the plotted values are colorized in a rainbow fashion this change will probably have a worse view, in that case I believe a string representation of the axis will be needed, but that's for another patch.

Reviewers: #kde_edu, apol, tcanabrava

Reviewed By: #kde_edu, tcanabrava

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D18304