Use Config mode for finding Eigen3
In [1] the build system was changed to use Eigen3's imported targets
provided upstream. These are only available with the cmake config
files, not the FindEigen3 module (also upstream). With the basic
signature of find_package() it searches in Module mode first, finding
Eigen3 via FindEigen3.cmake and erroring out afterwards because
Eigen3::Eigen is an unknown target.
Forcing find_package to use the Config mode fixes that.

[1] 7f3d443c206c599fd044d5ba485f2f9a2fb482c8