Analitza6Config: Add Qt6Core5Compat to dependencies
It is part of the public link interface, so cmake should make
sure it is available.
I suppose its absence from the cmake config file, is the reason it
was added to kalgebra [1], which doesn't really seem to use it
directly itself.
While at it, also use the right variable with the minimum Qt version.

[1] https://invent.kde.org/education/kalgebra/-/commit/137d730fbde23582635573a880ed208c0e2d981c